subscription_intelligence_mcp/requirements.txt
streamlit
rich
requests
python-dotenv
pydantic
matplotlib
